requirement.txt
# Core dependencies for BP Predictor
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0

# LLM client dependencies
requests>=2.28.0
openai>=0.27.0

# Optional: for other features
flask>=2.0.0
python-dotenv>=0.19.0

# Optional: for vision/rag features (if needed)
torch>=1.13.0
torchvision>=0.14.0
Pillow>=9.0.0
faiss-cpu>=1.7.0
llama-index>=0.5.0
langchain>=0.0.100
sentence-transformers>=2.2.0
